M07L05CalculatorUnitTestDemoApp/M07L05CalculationLibrary/Calculations.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace M07L05CalculationLibrary
{
    public class Calculations
    {
        public double Add(double a, double b)
        {
            return a + b;
        }

        public double Subtract(double a, double b)
        {
            return a - b;
        }

        public double Multiply(double a, double b)
        {
            return a * b;
        }

        public double Divide(double a, double b)
        {
            double output = 0;
            if( b != 0)
            {
                output = a / b;
            }
            return output;
        }
    }
}
